+++ example/test_driver/h264_e2e.dart
@@ -1,0 +1,11 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:e2e/e2e.dart';
+import 'package:h264/h264.dart';
+
+void main() {
+  E2EWidgetsFlutterBinding.ensureInitialized();
+
+  testWidgets('convert a h264 file', (WidgetTester tester) async {
+    // TODO
+  });
+}